ref: checkoutCard.tsx and fix review
