+++ components/domains/steps/checkoutCard.tsx
@@ -3,19 +3,15 @@ import React, {
   useCallback,
   useContext,
   useEffect,
-  useMemo,
   useState,
 } from "react";
 import styles from "../../../styles/components/registerV3.module.css";
 import { FormContext } from "@/context/FormProvider";
 import {
-  AutoRenewalContracts,
   CurrencyType,
-  ERC20Contract,
   FormType,
   NotificationType,
   TransactionType,
-  swissVatRate,
 } from "@/utils/constants";
 import { useAccount, useContractWrite } from "@starknet-react/core";
 import Button from "@/components/UI/button";
@@ -25,41 +21,19 @@ import useAllowanceCheck from "@/hooks/useAllowanceCheck";
 import useBalances from "@/hooks/useBalances";
 import { Divider } from "@mui/material";
 import { Call } from "starknet";
-import {
-  getAutoRenewAllowance,
-  getDomainPriceAltcoin,
-  getPriceForDuration,
-  getTokenQuote,
-  getYearlyPrice,
-  smartCurrencyChoosing,
-} from "@/utils/altcoinService";
-import {
-  getApprovalAmount,
-  getManyDomainsPriceWei,
-} from "@/utils/priceService";
-import {
-  applyRateToBigInt,
-  hexToDecimal,
-  toUint256,
-} from "@/utils/feltService";
 import RegisterCheckboxes from "../registerCheckboxes";
-import {
-  formatHexString,
-  getDomainWithStark,
-  selectedDomainsToArray,
-  selectedDomainsToEncodedArray,
-} from "@/utils/stringService";
+import { formatHexString, getDomainWithStark } from "@/utils/stringService";
 import UpsellCard from "./upsellCard";
-import registrationCalls from "@/utils/callData/registrationCalls";
-import { utils } from "starknetid.js";
-import autoRenewalCalls from "@/utils/callData/autoRenewalCalls";
-import { useDomainFromAddress } from "@/hooks/naming";
-import identityChangeCalls from "@/utils/callData/identityChangeCalls";
 import posthog from "posthog-js";
 import { useRouter } from "next/router";
 import { formatDomainData } from "@/utils/cacheDomainData";
 import ReduceDuration from "./reduceDuration";
 import Notification from "@/components/UI/notification";
+import { useCurrencyManagement } from "@/hooks/checkout/useCurrencyManagement";
+import { usePriceManagement } from "@/hooks/checkout/usePriceManagement";
+import { useCheckoutState } from "@/hooks/checkout/useCheckoutState";
+import { useRegisterTxPrep } from "@/hooks/checkout/useRegisterTxPrep";
+import { useRenewalTxPrep } from "@/hooks/checkout/useRenewalTxPrep";
 
 type CheckoutCardProps = {
   type: FormType;
@@ -75,587 +49,106 @@ const CheckoutCard: FunctionComponent<CheckoutCardProps> = ({
   const router = useRouter();
   const { account, address } = useAccount();
   const { formState, updateFormState, clearForm } = useContext(FormContext);
-  const [dailyPriceInEth, setDailyPriceInEth] = useState<bigint>(BigInt(0)); // Daily price in ETH
-  const [discountedPriceInEth, setDiscountedPriceInEth] = useState<bigint>(
-    BigInt(0)
-  ); // Daily discounted price in ETH
-  const [price, setPrice] = useState<bigint>(BigInt(0)); // Total price to pay in displayedCurrency
-  const [discountedPrice, setDiscountedPrice] = useState<bigint>(BigInt(0)); // Total discounted price to pay in displayedCurrency
-
-  const [maxPriceRange, setMaxPriceRange] = useState<bigint>(BigInt(0)); // max price range for the displayedCurrency that will be spent on yearly subscription
-  const [quoteData, setQuoteData] = useState<QuoteQueryData | null>(null); // null if in ETH
-  const [salesTaxAmount, setSalesTaxAmount] = useState<bigint>(BigInt(0));
-  const [invalidBalance, setInvalidBalance] = useState<boolean>(false);
-  const [renewalBox, setRenewalBox] = useState<boolean>(true);
-  const [termsBox, setTermsBox] = useState<boolean>(false);
-  const mainDomain = useDomainFromAddress(address ?? "");
-  const [mainDomainBox, setMainDomainBox] = useState<boolean>(true);
-  const [sponsor, setSponsor] = useState<string>("0x0");
-  const [hasUserSelectedOffer, setHasUserSelectedOffer] =
-    useState<boolean>(false);
-  const [displayedCurrency, setDisplayedCurrency] = useState<CurrencyType>(
-    CurrencyType.ETH
-  );
-  const [loadingPrice, setLoadingPrice] = useState<boolean>(false);
-  const [reloadingPrice, setReloadingPrice] = useState<boolean>(false); // used to know if the user changes the currency
-  const [hasReverseAddressRecord, setHasReverseAddressRecord] =
-    useState<boolean>(false);
-  const [domainsMinting, setDomainsMinting] =
-    useState<Record<string, boolean>>();
   const { addTransaction } = useNotificationManager();
-  const allowanceStatus = useAllowanceCheck(displayedCurrency, address);
   const tokenBalances = useBalances(address); // fetch the user balances for all whitelisted tokens
-  const [hasChosenCurrency, setHasChosenCurrency] = useState<boolean>(false);
   const [callData, setCallData] = useState<Call[]>([]);
-  const [tokenIdRedirect, setTokenIdRedirect] = useState<string>("0x0");
   const { writeAsync: execute, data: checkoutData } = useContractWrite({
     calls: callData,
   });
-  const [reducedDuration, setReducedDuration] = useState<number>(0); // reduced duration for the user to buy the domain
-  const [nonSubscribedDomains, setNonSubscribedDomains] = useState<string[]>();
-  const [currencyError, setCurrencyError] = useState<boolean>(false);
 
-  const hasMainDomain = useMemo(() => {
-    if (!mainDomain || !mainDomain.domain) return false;
-    return mainDomain.domain.endsWith(".stark");
-  }, [mainDomain]);
+  const {
+    termsBox,
+    renewalBox,
+    mainDomainBox,
+    hasUserSelectedOffer,
+    hasReverseAddressRecord,
+    domainsMinting,
+    reducedDuration,
+    nonSubscribedDomains,
+    sponsor,
+    hasMainDomain,
+    setHasUserSelectedOffer,
+    setDomainsMinting,
+    setReducedDuration,
+    onChangeTermsBox,
+    onChangeRenewalBox,
+    onChangeMainDomainBox,
+  } = useCheckoutState(type, address);
 
-  // Fetch token quote
-  useEffect(() => {
-    const fetchQuote = async () => {
-      if (displayedCurrency === CurrencyType.ETH) return;
-      const data = await getTokenQuote(ERC20Contract[displayedCurrency]);
-      if (data) {
-        setQuoteData(data);
-        setCurrencyError(false);
-      } else {
-        setDisplayedCurrency(CurrencyType.ETH);
-        setCurrencyError(true);
-      }
-    };
+  const {
+    displayedCurrency,
+    quoteData,
+    currencyError,
+    hasChosenCurrency,
+    onCurrencySwitch,
+    setHasChosenCurrency,
+    setCurrencyError,
+  } = useCurrencyManagement();
 
-    // refetch new quote if the timestamp from quote is expired
-    const scheduleRefetch = () => {
-      const now = parseInt((new Date().getTime() / 1000).toFixed(0));
-      const timeLimit = now - 60; // 60 seconds
-      // Check if we need to refetch
-      if (!quoteData || displayedCurrency === CurrencyType.ETH) {
-        setQuoteData(null);
-        // we don't need to check for quote until displayedCurrency is updated
-        return;
-      }
-
-      if (quoteData.max_quote_validity <= timeLimit) {
-        fetchQuote();
-      }
-
-      // Calculate the time until the next validity check
-      const timeUntilNextCheck = quoteData.max_quote_validity - timeLimit;
-      setTimeout(scheduleRefetch, Math.max(15000, timeUntilNextCheck * 100));
-    };
-
-    // Initial fetch
-    fetchQuote();
-    // Start the refetch scheduling
-    scheduleRefetch();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [displayedCurrency, dailyPriceInEth]); // We don't add quoteData because it would create an infinite loop
-
-  useEffect(() => {
-    // get the price for domains for a year
-    if (!formState.selectedDomains) return;
-
-    const _dailyPriceInEth = getManyDomainsPriceWei(
-      formState.selectedDomains,
-      1
-    );
-    setDailyPriceInEth(_dailyPriceInEth);
-    setDiscountedPriceInEth(
-      _dailyPriceInEth * BigInt(discount.paidDurationInDays)
-    );
-  }, [formState.selectedDomains, discount.paidDurationInDays]);
-
-  useEffect(() => {
-    const discountedPrice =
-      dailyPriceInEth * BigInt(discount.paidDurationInDays);
-
-    // Update discountedPrice based on isUpselled selected or not
-    if (formState.isUpselled) {
-      if (displayedCurrency === CurrencyType.ETH) {
-        setDiscountedPrice(discountedPrice);
-      } else if (quoteData) {
-        const priceInAltcoin = getDomainPriceAltcoin(
-          quoteData?.quote as string,
-          discountedPrice
-        );
-        setDiscountedPrice(priceInAltcoin);
-      }
-    } else setDiscountedPrice(BigInt(0));
-  }, [
+  const {
     dailyPriceInEth,
-    formState.isUpselled,
+    discountedPriceInEth,
+    price,
+    discountedPrice,
+    loadingPrice,
+    setLoadingPrice,
+    setReloadingPrice,
+    salesTaxAmount,
+    maxPriceRange,
+    invalidBalance,
+  } = usePriceManagement(
+    formState,
     displayedCurrency,
     quoteData,
     discount,
-  ]);
-
-  // we choose the currency based on the user balances
-  useEffect(() => {
-    if (
-      tokenBalances &&
-      Object.keys(tokenBalances).length > 0 &&
-      !hasChosenCurrency &&
-      dailyPriceInEth
-    ) {
-      const domainPrice = formState.isUpselled
-        ? dailyPriceInEth * BigInt(discount.paidDurationInDays)
-        : dailyPriceInEth * BigInt(formState.durationInYears * 365);
-      smartCurrencyChoosing(tokenBalances, domainPrice).then((currency) => {
-        onCurrencySwitch(currency);
-        setHasChosenCurrency(true);
-      });
-    }
-  }, [
     tokenBalances,
-    dailyPriceInEth,
-    formState.isUpselled,
     hasChosenCurrency,
-    formState.durationInYears,
-    discount.paidDurationInDays,
-  ]);
+    onCurrencySwitch,
+    setHasChosenCurrency,
+    updateFormState
+  );
+  const allowanceStatus = useAllowanceCheck(displayedCurrency, address);
 
-  // we ensure user has enough balance of the token selected
-  useEffect(() => {
-    if (
-      tokenBalances &&
-      price &&
-      displayedCurrency &&
-      !loadingPrice &&
-      !reloadingPrice
-    ) {
-      const tokenBalance = tokenBalances[displayedCurrency];
-      if (!tokenBalance) return;
-      const _price = formState.isUpselled ? discountedPrice : price;
-      if (tokenBalance && BigInt(tokenBalance) >= BigInt(_price))
-        setInvalidBalance(false);
-      else setInvalidBalance(true);
-    }
-  }, [
+  const {
+    callData: registerCallData,
+    tokenIdRedirect: registerTokenIdRedirect,
+  } = useRegisterTxPrep(
+    formState,
+    displayedCurrency,
+    quoteData,
     price,
     discountedPrice,
-    formState.isUpselled,
-    displayedCurrency,
-    tokenBalances,
-    loadingPrice,
-    reloadingPrice,
-  ]);
-
-  useEffect(() => {
-    const referralData = localStorage.getItem("referralData");
-    if (referralData) {
-      const data = JSON.parse(referralData);
-      if (data.sponsor && data?.expiry >= new Date().getTime()) {
-        setSponsor(data.sponsor);
-      } else {
-        setSponsor("0");
-      }
-    } else {
-      setSponsor("0");
-    }
-  }, [address, formState.selectedDomains]);
-
-  useEffect(() => {
-    const _price = formState.isUpselled ? discountedPrice : price;
-    if (!formState.needMetadata && _price) {
-      setSalesTaxAmount(applyRateToBigInt(_price, formState.salesTaxRate));
-    } else {
-      if (formState.isSwissResident) {
-        updateFormState({ salesTaxRate: swissVatRate });
-        setSalesTaxAmount(applyRateToBigInt(_price, swissVatRate));
-      } else {
-        updateFormState({ salesTaxRate: 0 });
-        setSalesTaxAmount(BigInt(0));
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [
-    formState.isSwissResident,
-    price,
-    formState.isUpselled,
-    discountedPrice,
-    formState.needMetadata,
-    formState.salesTaxRate,
-  ]); // Don't call updateFromState on every render
-
-  // if dailyPriceInEth or quoteData have changed, we update the price in altcoin
-  useEffect(() => {
-    if (!dailyPriceInEth) return;
-    const _price = getManyDomainsPriceWei(
-      formState.selectedDomains,
-      formState.isUpselled
-        ? discount.durationInDays
-        : formState.durationInYears * 365
-    );
-    if (displayedCurrency === CurrencyType.ETH) {
-      setPrice(_price);
-    } else if (quoteData) {
-      setPrice(getDomainPriceAltcoin(quoteData.quote, _price));
-    }
-    setLoadingPrice(false);
-    if (reloadingPrice) setReloadingPrice(false);
-  }, [
-    dailyPriceInEth,
-    quoteData,
-    displayedCurrency,
-    formState.isUpselled,
-    formState.durationInYears,
-    discount.durationInDays,
-    reloadingPrice,
-    formState.selectedDomains,
-  ]);
-
-  useEffect(() => {
-    if (displayedCurrency !== CurrencyType.ETH && !quoteData) return;
-    const limitPrice = getAutoRenewAllowance(
-      displayedCurrency,
-      formState.salesTaxRate,
-      getYearlyPrice(
-        Object.keys(formState.selectedDomains)[0],
-        displayedCurrency,
-        quoteData?.quote
-      )
-    );
-    setMaxPriceRange(limitPrice);
-  }, [
-    displayedCurrency,
-    formState.selectedDomains,
-    quoteData,
-    formState.salesTaxRate,
-  ]);
-
-  useEffect(() => {
-    if (!address) return;
-    fetch(`${process.env.NEXT_PUBLIC_SERVER_LINK}/addr_has_rev?addr=${address}`)
-      .then((response) => response.json())
-      .then((reverseAddressData) => {
-        setHasReverseAddressRecord(reverseAddressData.has_rev);
-      });
-  }, [address]);
-
-  // we get the list of domains that do not have a autorenewal already enabled
-  useEffect(() => {
-    if (type !== FormType.RENEW) return;
-    if (address) {
-      fetch(
-        `${process.env.NEXT_PUBLIC_SERVER_LINK}/renewal/get_non_subscribed_domains?addr=${address}`
-      )
-        .then((response) => response.json())
-        .then((data) => {
-          setNonSubscribedDomains(data);
-        });
-    }
-  }, [address, formState.selectedDomains, renewalBox, type]);
-
-  // Set Register Multicall
-  useEffect(() => {
-    if (
-      (displayedCurrency !== CurrencyType.ETH && !quoteData) ||
-      type !== FormType.REGISTER
-    )
-      return;
-    // Variables
-    const newTokenId: number = Math.floor(Math.random() * 1000000000000);
-    const txMetadataHash = `0x${formState.metadataHash}` as HexString;
-    const encodedDomain = utils
-      .encodeDomain(Object.keys(formState.selectedDomains)[0])
-      .map((element) => element.toString())[0];
-    const finalDurationInDays = formState.isUpselled
-      ? discount.durationInDays
-      : formState.durationInYears * 365;
-    const priceToPay = formState.isUpselled ? discountedPrice : price;
-
-    // Common calls
-    const calls = [
-      registrationCalls.approve(priceToPay, ERC20Contract[displayedCurrency]),
-    ];
-
-    if (displayedCurrency === CurrencyType.ETH) {
-      calls.push(
-        registrationCalls.buy(
-          encodedDomain,
-          formState.tokenId === 0 ? newTokenId : formState.tokenId,
-          sponsor,
-          finalDurationInDays,
-          txMetadataHash,
-          formState.isUpselled ? discount.discountId : "0"
-        )
-      );
-    } else {
-      calls.push(
-        registrationCalls.altcoinBuy(
-          encodedDomain,
-          formState.tokenId === 0 ? newTokenId : formState.tokenId,
-          sponsor,
-          finalDurationInDays,
-          txMetadataHash,
-          ERC20Contract[displayedCurrency],
-          quoteData as QuoteQueryData,
-          formState.isUpselled ? discount.discountId : "0"
-        )
-      );
-    }
-
-    // If the user is a Swiss resident, we add the sales tax
-    if (formState.salesTaxRate) {
-      calls.unshift(
-        registrationCalls.vatTransfer(
-          salesTaxAmount,
-          ERC20Contract[displayedCurrency]
-        )
-      ); // IMPORTANT: We use unshift to put the call at the beginning of the array
-    }
-
-    // If the user choose to mint a new identity
-    let tokenIdToUse = formState.tokenId;
-    if (formState.tokenId === 0) {
-      calls.unshift(registrationCalls.mint(newTokenId)); // IMPORTANT: We use unshift to put the call at the beginning of the array
-      tokenIdToUse = newTokenId;
-    }
-    setTokenIdRedirect(tokenIdToUse.toString());
-
-    // If the user does not have a main domain or has checked the mainDomain box
-    if (!hasMainDomain || mainDomainBox) {
-      if (hasReverseAddressRecord)
-        calls.push(registrationCalls.resetAddrToDomain());
-      calls.push(registrationCalls.mainId(tokenIdToUse));
-    }
-
-    if (renewalBox) {
-      if (allowanceStatus.needsAllowance) {
-        const amountToApprove = getApprovalAmount(
-          price,
-          salesTaxAmount,
-          discountedPrice
-            ? Number(
-                BigInt(discount.durationInDays) / BigInt(finalDurationInDays)
-              )
-            : formState.durationInYears,
-          allowanceStatus.currentAllowance
-        );
-
-        calls.push(
-          autoRenewalCalls.approve(
-            ERC20Contract[displayedCurrency],
-            AutoRenewalContracts[displayedCurrency],
-            amountToApprove
-          )
-        );
-      }
-
-      const allowance = getAutoRenewAllowance(
-        displayedCurrency,
-        formState.salesTaxRate,
-        getYearlyPrice(
-          Object.keys(formState.selectedDomains)[0],
-          displayedCurrency,
-          quoteData?.quote
-        )
-      );
-      calls.push(
-        autoRenewalCalls.enableRenewal(
-          AutoRenewalContracts[displayedCurrency],
-          encodedDomain,
-          allowance,
-          `0x${formState.metadataHash}`
-        )
-      );
-    }
-
-    // if the user has selected a profile picture
-    if (formState.selectedPfp) {
-      const nftData = formState.selectedPfp;
-      const nft_id = toUint256(nftData.token_id);
-      calls.push(
-        identityChangeCalls.updateProfilePicture(
-          hexToDecimal(nftData.contract_address),
-          nft_id.low,
-          nft_id.high,
-          tokenIdToUse.toString()
-        )
-      );
-    }
-
-    // Merge and set the call data
-    setCallData(calls);
-  }, [
-    formState.tokenId,
-    formState.isUpselled,
-    formState.durationInYears,
-    price,
-    formState.selectedDomains,
-    hasMainDomain,
-    address,
-    sponsor,
-    formState.metadataHash,
-    formState.salesTaxRate,
-    renewalBox,
-    mainDomainBox,
     salesTaxAmount,
-    allowanceStatus.needsAllowance,
-    allowanceStatus.currentAllowance,
-    quoteData,
-    displayedCurrency,
-    formState.selectedPfp,
-    type,
-    discount.durationInDays,
-    discount.discountId,
-    discountedPrice,
+    sponsor,
+    hasMainDomain,
+    mainDomainBox,
     hasReverseAddressRecord,
-  ]);
-
-  // Set Renewal Multicall
-  useEffect(() => {
-    if (type !== FormType.RENEW) return;
-    if (displayedCurrency !== CurrencyType.ETH && !quoteData) return;
-
-    // Variables
-    const finalDurationInDays = formState.isUpselled
-      ? discount.durationInDays
-      : formState.durationInYears * 365;
-    const priceToPay = formState.isUpselled ? discountedPrice : price;
-    const txMetadataHash = `0x${formState.metadataHash}` as HexString;
-
-    // Common calls
-    const calls = [
-      registrationCalls.approve(priceToPay, ERC20Contract[displayedCurrency]),
-    ];
-
-    if (displayedCurrency === CurrencyType.ETH) {
-      calls.push(
-        ...registrationCalls.multiCallRenewal(
-          selectedDomainsToEncodedArray(formState.selectedDomains),
-          finalDurationInDays,
-          txMetadataHash,
-          sponsor,
-          formState.isUpselled ? discount.discountId : "0"
-        )
-      );
-    } else {
-      calls.push(
-        ...registrationCalls.multiCallRenewalAltcoin(
-          selectedDomainsToEncodedArray(formState.selectedDomains),
-          finalDurationInDays,
-          txMetadataHash,
-          ERC20Contract[displayedCurrency],
-          quoteData as QuoteQueryData,
-          sponsor,
-          formState.isUpselled ? discount.discountId : "0"
-        )
-      );
-    }
-
-    // If the user is a Swiss resident, we add the sales tax
-    if (formState.salesTaxRate) {
-      calls.unshift(
-        registrationCalls.vatTransfer(
-          salesTaxAmount,
-          ERC20Contract[displayedCurrency]
-        )
-      ); // IMPORTANT: We use unshift to put the call at the beginning of the array
-    }
-
-    // If the user has toggled autorenewal
-    if (renewalBox) {
-      if (allowanceStatus.needsAllowance) {
-        const amountToApprove = getApprovalAmount(
-          price,
-          salesTaxAmount,
-          discountedPrice
-            ? Number(BigInt(discount.durationInDays) / BigInt(365))
-            : formState.durationInYears,
-          allowanceStatus.currentAllowance
-        );
-
-        calls.push(
-          autoRenewalCalls.approve(
-            ERC20Contract[displayedCurrency],
-            AutoRenewalContracts[displayedCurrency],
-            amountToApprove
-          )
-        );
-      }
-
-      selectedDomainsToArray(formState.selectedDomains).map((domain) => {
-        // we enable renewal only for the domains that are not already subscribed
-        if (nonSubscribedDomains?.includes(domain)) {
-          const encodedDomain = utils
-            .encodeDomain(domain)
-            .map((element) => element.toString())[0];
-
-          const domainPrice = getYearlyPrice(
-            domain,
-            displayedCurrency,
-            quoteData?.quote
-          );
-          const allowance = getAutoRenewAllowance(
-            displayedCurrency,
-            formState.salesTaxRate,
-            domainPrice
-          );
-
-          calls.push(
-            autoRenewalCalls.enableRenewal(
-              AutoRenewalContracts[displayedCurrency],
-              encodedDomain,
-              allowance,
-              txMetadataHash
-            )
-          );
-        }
-      });
-    }
-
-    // if the user has selected a profile picture
-    if (formState.selectedPfp) {
-      const nftData = formState.selectedPfp;
-      const nft_id = toUint256(nftData.token_id);
-      calls.push(
-        identityChangeCalls.updateProfilePicture(
-          hexToDecimal(nftData.contract_address),
-          nft_id.low,
-          nft_id.high,
-          formState.tokenId.toString()
-        )
-      );
-    }
-
-    // Merge and set the call data
-    setCallData(calls);
-  }, [
-    formState.tokenId,
-    formState.isUpselled,
-    formState.durationInYears,
-    price,
-    formState.selectedDomains,
-    hasMainDomain,
-    address,
-    sponsor,
-    formState.metadataHash,
-    formState.salesTaxRate,
+    discount,
     renewalBox,
-    mainDomainBox,
-    salesTaxAmount,
-    allowanceStatus.needsAllowance,
-    allowanceStatus.currentAllowance,
-    quoteData,
+    allowanceStatus
+  );
+
+  const { callData: renewalCallData } = useRenewalTxPrep(
+    formState,
     displayedCurrency,
-    formState.selectedPfp,
-    type,
-    discount.durationInDays,
-    discount.discountId,
+    quoteData,
+    price,
     discountedPrice,
-    nonSubscribedDomains,
-  ]);
+    salesTaxAmount,
+    sponsor,
+    discount,
+    renewalBox,
+    allowanceStatus,
+    nonSubscribedDomains
+  );
+
+  useEffect(() => {
+    if (type === FormType.REGISTER) {
+      setCallData(registerCallData);
+    } else if (type === FormType.RENEW) {
+      setCallData(renewalCallData);
+    }
+  }, [type, registerCallData, renewalCallData]);
 
   // on execute transaction,
   useEffect(() => {
@@ -711,7 +204,7 @@ const CheckoutCard: FunctionComponent<CheckoutCardProps> = ({
     // if registration, store domain data in local storage to use it until it's indexed
     if (type === FormType.REGISTER) {
       formatDomainData(
-        tokenIdRedirect,
+        registerTokenIdRedirect,
         formatHexString(address as string),
         getDomainWithStark(Object.keys(formState.selectedDomains)[0]),
         formState.isUpselled
@@ -727,17 +220,17 @@ const CheckoutCard: FunctionComponent<CheckoutCardProps> = ({
 
     // Redirect to confirmation page
     if (type === FormType.REGISTER)
-      router.push(`/confirmation?tokenId=${tokenIdRedirect}`);
+      router.push(`/confirmation?tokenId=${registerTokenIdRedirect}`);
     else router.push(`/confirmation`);
 
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [checkoutData]); // We only need registerData here because we don't want to send the metadata twice (we send it once the tx is sent)
 
-  const onCurrencySwitch = (type: CurrencyType) => {
+  const handleCurrencySwitch = (type: CurrencyType) => {
     setReloadingPrice(true);
     if (type !== CurrencyType.ETH) setLoadingPrice(true);
     setReducedDuration(0);
-    setDisplayedCurrency(type);
+    onCurrencySwitch(type);
     setHasUserSelectedOffer(false);
   };
 
@@ -747,47 +240,6 @@ const CheckoutCard: FunctionComponent<CheckoutCardProps> = ({
     },
     [updateFormState]
   );
-
-  useEffect(() => {
-    const durationInYears = formState.durationInYears;
-    if (!invalidBalance || !dailyPriceInEth) {
-      if (!reducedDuration) setReducedDuration(0);
-      return;
-    }
-    let found = false;
-    for (
-      let newDurationInYears = durationInYears - 1;
-      newDurationInYears > 0;
-      newDurationInYears--
-    ) {
-      const newPriceInEth = getPriceForDuration(
-        dailyPriceInEth,
-        newDurationInYears
-      );
-      let newPrice = newPriceInEth;
-      if (displayedCurrency !== CurrencyType.ETH && quoteData)
-        newPrice = getDomainPriceAltcoin(quoteData.quote, newPriceInEth);
-      const balance = tokenBalances[displayedCurrency];
-      if (!balance) continue;
-      if (BigInt(balance) >= BigInt(newPrice)) {
-        if (reducedDuration !== newDurationInYears)
-          setReducedDuration(newDurationInYears);
-        found = true;
-        break;
-      }
-    }
-    if (!found && reducedDuration) setReducedDuration(0);
-  }, [
-    formState.durationInYears,
-    invalidBalance,
-    discount.durationInDays,
-    dailyPriceInEth,
-    formState.isUpselled,
-    tokenBalances,
-    displayedCurrency,
-    quoteData,
-    reducedDuration,
-  ]);
 
   return (
     <>
@@ -829,7 +281,7 @@ const CheckoutCard: FunctionComponent<CheckoutCardProps> = ({
             salesTaxRate={formState.salesTaxRate}
             isSwissResident={formState.isSwissResident}
             displayedCurrency={displayedCurrency}
-            onCurrencySwitch={onCurrencySwitch}
+            onCurrencySwitch={handleCurrencySwitch}
             loadingPrice={loadingPrice}
             isUpselled={formState.isUpselled}
             discountedPrice={discountedPrice}
@@ -838,13 +290,13 @@ const CheckoutCard: FunctionComponent<CheckoutCardProps> = ({
           <Divider className="w-full" />
           <div className={styles.checkoutSummary}>
             <RegisterCheckboxes
-              onChangeTermsBox={() => setTermsBox(!termsBox)}
+              onChangeTermsBox={onChangeTermsBox}
               termsBox={termsBox}
-              onChangeRenewalBox={() => setRenewalBox(!renewalBox)}
+              onChangeRenewalBox={onChangeRenewalBox}
               renewalBox={renewalBox}
               showMainDomainBox={type === FormType.REGISTER && hasMainDomain}
               mainDomainBox={mainDomainBox}
-              onChangeMainDomainBox={() => setMainDomainBox(!mainDomainBox)}
+              onChangeMainDomainBox={onChangeMainDomainBox}
               domain={getDomainWithStark(
                 Object.keys(formState.selectedDomains)[0]
               )}
