+++ hooks/useNeedAllowances.tsx
@@ -1,7 +1,7 @@
 import { useContractRead } from "@starknet-react/core";
 import { useMulticallContract } from "./contracts";
 import { Abi, BlockTag, CairoCustomEnum, Call, RawArgs, hash } from "starknet";
-import { useEffect, useState } from "react";
+import { useEffect, useState, useMemo } from "react";
 import {
   ERC20Contract,
   CurrencyType,
@@ -9,15 +9,16 @@ import {
 } from "../utils/constants";
 import { isApprovalInfinite } from "@/utils/priceService";
 
-export default function useNeedsAllowances(
-  address?: string
-): TokenNeedsAllowance {
-  const initialAllowances: TokenNeedsAllowance = Object.values(
-    CurrencyType
-  ).reduce((acc, currency) => {
+const createInitialAllowances = (): TokenNeedsAllowance =>
+  Object.values(CurrencyType).reduce((acc, currency) => {
     acc[currency] = { needsAllowance: false, currentAllowance: BigInt(0) };
     return acc;
   }, {} as TokenNeedsAllowance);
+
+export default function useNeedsAllowances(
+  address?: string
+): TokenNeedsAllowance {
+  const initialAllowances = useMemo(() => createInitialAllowances(), []);
   const [needsAllowances, setNeedsAllowances] =
     useState<TokenNeedsAllowance>(initialAllowances);
   const [callData, setCallData] = useState<Call[]>([]);
