+++ hooks/useNeedSubscription.tsx
@@ -1,5 +1,6 @@
 import { useState, useEffect } from "react";
 import { hexToDecimal } from "@/utils/feltService";
+import { processSubscriptionData } from "@/utils/subscriptionService";
 
 type UseNeedSubscriptionResult = {
   needSubscription: NeedSubscription;
@@ -25,15 +26,7 @@ export default function useNeedSubscription(
       )
         .then((response) => response.json())
         .then((data: SubscriptionInfos) => {
-          const processedData: NeedSubscription = {};
-          Object.entries(data).forEach(([domain, subscriptions]) => {
-            if (subscriptions.eth_subscriptions === null) {
-              processedData[domain] = {
-                ETH: { needsAllowance: true, currentAllowance: BigInt(0) },
-                STRK: { needsAllowance: true, currentAllowance: BigInt(0) },
-              };
-            }
-          });
+          const processedData = processSubscriptionData(data);
           setNeedSubscription(processedData);
           setIsLoading(false);
         });
